requirements.txt
aiomysql==0.3.2
aiosqlite==0.21.0
bcrypt==5.0.0
blinker==1.9.0
cachelib==0.13.0
click==8.3.0
colorama==0.4.6
flask==3.1.2
flask-session==0.8.0
iso8601==2.1.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
msgspec==0.20.0
mysql-connector-python==9.5.0
pymysql==1.1.2
pypika-tortoise==0.6.3
python-dotenv==1.2.1
pytz==2025.2
tortoise-orm==0.25.1
typing-extensions==4.15.0
werkzeug==3.1.3
